Registration done, but without Encrypt

+++ src/main/java/org/myproject/springmvc/dao/UserDAO.java
@@ -5,6 +5,6 @@ import org.myproject.springmvc.model.User;
 public interface UserDAO {
 	
 	User findByUsername(String username);
-	int save(User user);
+	Long save(User user);
 
 }
